remove history saving + rename
